requirements.txt
astroid==2.2.5
Django==2.2.3
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
pep8==1.7.1
psycopg2==2.8.3
PyJWT==1.7.1
pylint==2.3.1
python-dotenv==0.10.3
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
